Moved implementations of buildToEntityMapFunc() into separate classes
