+++ annotation-processor/src/main/kotlin/pl/touk/krush/KrushAnnotationProcessor.kt
@@ -3,7 +3,9 @@ package pl.touk.krush
 import com.squareup.kotlinpoet.metadata.KotlinPoetMetadataPreview
 import pl.touk.krush.env.EnvironmentBuilder
 import pl.touk.krush.model.EntityGraphBuilder
+import pl.touk.krush.source.CopiedReferencesMappingsGenerator
 import pl.touk.krush.source.MappingsGenerator
+import pl.touk.krush.source.RealReferencesMappingsGenerator
 import pl.touk.krush.source.TablesGenerator
 import java.io.File
 import javax.annotation.processing.AbstractProcessor
@@ -50,7 +52,14 @@ class KrushAnnotationProcessor : AbstractProcessor() {
         if (graphs.isEmpty()) return false
 
         try {
-            val generators = listOf(TablesGenerator(), MappingsGenerator(generateRealReferences = generateRealReferences))
+            val mappingsGenerator: MappingsGenerator
+            if (generateRealReferences) {
+                mappingsGenerator = RealReferencesMappingsGenerator()
+            } else {
+                mappingsGenerator = CopiedReferencesMappingsGenerator()
+            }
+
+            val generators = listOf(TablesGenerator(), mappingsGenerator)
             generators.forEach { generator ->
                 graphs.entries.forEach { (packageName, graph) ->
                     val poetFile = generator.generate(graph, graphs, packageName, envBuilder.buildTypeEnv())
